// libs/data-access/legacy/src/lib/legacy-application.dto.ts
export interface LegacyApplicationDto {
    id: string;

    citizenId?: string;
    citizen_id?: string;

    permitId?: string;
    permit_id?: string;

    region?: string;

    status?: string;
    application_status?: string;

    createdAt?: string;
    created_at?: string;

    updatedAt?: string;
    updated_at?: string;
}